+++ internal/server/dto/request/mother_service.go
@@ -2,8 +2,8 @@ package request
 
 type MotherService struct {
 	Name                     string  `json:"name"`
-	ExceptionRate            float64 `json:"exception_rate"`
-	ResponseDelayRate        float64 `json:"response_delay_rate"`
+	ExceptionRate            int     `json:"exception_rate"`
+	ResponseDelayRate        int     `json:"response_delay_rate"`
 	ResponseDelayDuration    *int    `json:"response_delay_duration"`
 	RandomResponseDelayMin   *int    `json:"random_response_delay_min"`
 	RandomResponseDelayMax   *int    `json:"random_response_delay_max"`
